+++ packages/theme/src/client/layouts/Layout.ts
@@ -9,7 +9,7 @@ import HomePage from "@theme-hope/components/HomePage";
 import NormalPage from "@theme-hope/components/NormalPage";
 import PortfolioHome from "@theme-hope/components/PortfolioHome";
 import SkipLink from "@theme-hope/components/SkipLink";
-import { FadeSlideY } from "@theme-hope/components/transitions/index";
+import { FadeInUpTransition } from "@theme-hope/components/transitions/index";
 import {
   usePure,
   useThemeData,
@@ -39,7 +39,7 @@ export default defineComponent({
   }>,
 
   setup(_props, { slots }) {
-    const themeData = useThemeData();
+    const theme = useThemeData();
     const themeLocale = useThemeLocaleData();
     const page = usePageData();
     const frontmatter = usePageFrontmatter<ThemePageFrontmatter>();
@@ -49,7 +49,7 @@ export default defineComponent({
     const sidebarDisplay = computed(() =>
       __VP_BLOG__
         ? (themeLocale.value.blog?.sidebarDisplay ??
-          themeData.value.blog?.sidebarDisplay ??
+          theme.value.blog?.sidebarDisplay ??
           "mobile")
         : "none",
     );
@@ -66,7 +66,7 @@ export default defineComponent({
               ? h(PortfolioHome)
               : frontmatter.value.home
                 ? h(HomePage)
-                : h(isPure.value ? RenderDefault : FadeSlideY, () =>
+                : h(isPure.value ? RenderDefault : FadeInUpTransition, () =>
                     h(
                       NormalPage,
                       { key: page.value.path },