+++ packages/theme/src/client/modules/outlook/components/OutlookSettings.ts
@@ -5,7 +5,7 @@ import { ClientOnly } from "vuepress/client";
 import { usePure, useThemeData } from "@theme-hope/composables/index";
 import ColorMode from "@theme-hope/modules/outlook/components/ColorMode";
 import ThemeColor, {
-  enableThemeColor,
+  hasMultipleThemeColors,
 } from "@theme-hope/modules/outlook/components/ThemeColor";
 import ToggleFullScreen from "@theme-hope/modules/outlook/components/ToggleFullScreen";
 
@@ -13,16 +13,16 @@ export default defineComponent({
   name: "OutlookSettings",
 
   setup() {
-    const themeData = useThemeData();
+    const theme = useThemeData();
     const isPure = usePure();
 
     const enableFullScreen = computed(
-      () => !isPure.value && themeData.value.fullscreen,
+      () => !isPure.value && theme.value.fullscreen,
     );
 
     return (): VNode =>
       h(ClientOnly, () => [
-        enableThemeColor ? h(ThemeColor) : null,
+        hasMultipleThemeColors ? h(ThemeColor) : null,
         h(ColorMode),
         enableFullScreen.value ? h(ToggleFullScreen) : null,
       ]);